src/main/java/com/biperf/core/domain/enums/PromotionIssuanceType.java
/**
 * 
 */

package com.biperf.core.domain.enums;

import java.util.List;

/**
 * PromotionIssuanceType.
 * <p>
 * <b>Change History:</b><br>
 * <table border="1">
 * <tr>
 * <th>Author</th>
 * <th>Date</th>
 * <th>Version</th>
 * <th>Comments</th>
 * </tr>
 * <tr>
 * <td>asondgeroth</td>
 * <td>Oct 10, 2005</td>
 * <td>1.0</td>
 * <td>created</td>
 * </tr>
 * </table>
 * 
 *
 */
public class PromotionIssuanceType extends PickListItem
{

  /**
   * Asset name used in Content Manager
   */
  private static final String PICKLIST_ASSET = "picklist.promotion.issuancetype";

  public static final String ONLINE = "online";
  public static final String FILE_LOAD = "file_load";

  /**
   * This constructor is protected - only the PickListFactory class creates these instances.
   */
  protected PromotionIssuanceType()
  {
    super();
  }

  /**
   * Get the pick list from content manager.
   * 
   * @return List
   */
  public static List getList()
  {
    return getPickListFactory().getPickList( PromotionIssuanceType.class );
  }

  /**
   * Returns null if the code is null or invalid for this list.
   * 
   * @param code
   * @return PickListItem
   */
  public static PromotionIssuanceType lookup( String code )
  {
    return (PromotionIssuanceType)getPickListFactory().getPickListItem( PromotionIssuanceType.class, code );
  }

  /**
   * Overridden from
   * 
   * @see com.biperf.core.domain.enums.PickListItem#getPickListAssetCode()
   * @return PICKLIST_ASSET
   */
  public String getPickListAssetCode()
  {
    return PICKLIST_ASSET;
  }

}
